Todos os 22 testes passando
Correção no sw e srli

+++ memoria.h
@@ -11,4 +11,4 @@ int32_t lbu(uint32_t address, int32_t kte);
 int32_t lhu(uint32_t address, int32_t kte);
 
 void sb(uint32_t address, int32_t kte, int8_t dado);
-void sw(uint32_t address, int32_t kte, int8_t dado);
+void sw(uint32_t address, int32_t kte, int32_t dado);
